requirements.txt
google-generativeai
streamlit
pillow
ipython
opencv-python
rich
streamlit
Live
python-dotenv
streamlit-chat